Add, solve Day 7
This one was painful. I'm still not entirely happy about the interior
mutability of the Nodes; it feels like it may have been more elegant
to use a builder object with interior mutability which could have been
converted into Node objects which did not have that feature. Still,
at least it's finished now.

Note that the part 2 solution requires a tiny bit of observation
and math on the part of the user. I could have completed the solver,
but it was getting painful to work on this code.

One thing which definitely didn't help was the iteration time. Cargo
consistently timed out when the editor called it, spamming the editor
window; it took on average something like 7 seconds to compile, and
then over 2 to run in release mode. Not the longest thing in the world,
but ten times longer than any other exercise so far. I'm not so bothered
by the compile times, given that it was probably regenerating the parser
every time, but the runtime does kind of bother me. It shouldn't take
that long to assemble a tree and then perform basic recursive analysis
on it.

In short, this is inelegant and I'm not super happy with it, but it
does in fact work. Could be worse, I guess.