Add, solve Day 11
This was a fun one; I hadn't thought much about hex coordinate systems
in the past, but it was good to see how they worked. I'm also pretty pleased
with the efficiency of the solution I came up with; 18 ms isn't bad when
the work is doubled!

... though I'm probably going to go back and fix that now, because it really isn't ideal to
do everything twice when there's an obvious single-pass solution. It's just that I went to
the trouble to implement iter::Sum for HexPosition, so I wanted to at least use that, even
though after Part 2 it's smarter not to.

$ cargo check && cargo build --release && time target/release/day11
   Compiling day11 v0.1.0 (file:///mnt/d/Users/coriolinus/Documents/Projects/adventofcode.com/adventofcode-2017/day11)
    Finished dev [unoptimized + debuginfo] target(s) in 0.64 secs
   Compiling util v0.2.0 (file:///mnt/d/Users/coriolinus/Documents/Projects/adventofcode.com/adventofcode-2017/util)
   Compiling day11 v0.1.0 (file:///mnt/d/Users/coriolinus/Documents/Projects/adventofcode.com/adventofcode-2017/day11)
    Finished release [optimized] target(s) in 4.74 secs
Dist to origin: 720
Max dist to origin: 1485

real    0m0.018s
user    0m0.000s
sys     0m0.016s
$ wc -c input.txt
21533 input.txt