Add and solve day15
This isn't actually a ton of code: I had a non-wrapping, non-bigint
mul_mod implementation lying around which I copied into util in case
it's useful there in the future, and the rest is pretty simple generator
stuff, which just isn't that hard.

I can't decide if I could have done this in a more efficient manner.
On the one hand, it's the slowest test so far by a long shot:

$ cargo build --release && time (target/release/day15 > /dev/null)
    Finished release [optimized] target(s) in 0.0 secs

real    0m28.164s
user    0m28.109s
sys     0m0.031s

On the other hand, it's hard to parallelize because the state of the
generator always depends on its previous value. I suppose I could
use rayon to create a composite generator which computes the next
value of the a and b generators in separate threads, but I'm not entirely
convinced that the potential speedup is worth the development effort.

... might end up trying it anyway, just for fun. Not sure.