Replace mul_mod with standard math operations.
Turns out it's _substantially_ faster to just do the naive thing
and multiply the two values, then take the modulus, than to assume
that one of the multiplications will eventually overflow a u64.
I should have done the math to figure out if overflow was possible
before reaching for a tool which handled that case.

This is now much faster:
$ cargo build --release && time (target/release/day15 > /dev/null)
    Finished release [optimized] target(s) in 0.0 secs

real    0m1.589s
user    0m1.563s
sys     0m0.000s

No need to attempt to parallelize it after all.