add, complete day 21
turns out that on vacation it is _very hard_ to make time to do
this kind of thing; I find myself two days behind right now, and
I've been trying more or less continuously to make time for this.

It works and it's fast enough for now--it runs part 2 in a few
seconds--but there are a ton of copies throughout. Essentially
I gave up on writing an zero-copy elegant solution and focused
on just getting something out the door. Hopefully it's good enough.

Also hopefully day 22 will be quicker! This really took longer than
I'd have liked.