Add, solve day 23.
That one was rough. Got it working eventually, but decompiling
assembly is not my favorite thing.

Lots of extra files in this one: input.txt.parsed is my midway point
figuring out the structure of the input; further iteration on that
produced input.py. Finally, after verifying that input.py generated
the correct answer, I actually wrote the part2 solution in Rust.