Add, complete day24
This
took
_forever_.

I did start on the 24th, I think. I'm writing this commit message
one minute before midnight on the 30th. That wasn't all time
working on the code, of course, but the code itself was also
obstinate and uncooperative this time for some reason.

The real issue was in designing the iterator. Of course, in
retrospect, the iterator is the heart of the problem, but for
whatever reason it just didn't want to come together, and it
took three or four design iterations before it actually
worked correctly. Still, once it did, the rest was trivial.